ai-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
langchain
langchain-google-genai
langchain-community
langchain-chroma
chromadb
pypdf
google-genai
transformers
torch
sentencepiece
